+++ irodsadmin/management/commands/checksampleaccess.py
@@ -50,10 +50,20 @@ class Command(BaseCommand):
     def __init__(self):
         super().__init__()
         self.irods_backend = plugin_api.get_backend_api('omics_irods')
+        self.log_level = getattr(settings, 'LOGGING_LEVEL')
 
-    @classmethod
+    def _log_debug(self, msg: str):
+        """
+        Helper to only log/print debug messages with appropriate LOGGING_LEVEL
+        (prevents unwanted stdout printing of debug messages)
+
+        :param msg: Log message (string)
+        """
+        if self.log_level == 'DEBUG':
+            logger.debug(msg)
+
     def _check_access(
-        cls,
+        self,
         user_id: int,
         user_name: str,
         access_name: str,
@@ -71,6 +81,10 @@ class Command(BaseCommand):
         :param admin_id: Admin user ID (int)
         :param group_id: Project user group ID (int)
         """
+        self._log_debug(
+            f'Checking ACL: user_id={user_id}; user_name={user_name}; '
+            f'access_name={access_name}; path={path}'
+        )
         if user_id == admin_id and access_name != IRODS_ACCESS_OWN:
             logger.info(f'{CHECK_ACCESS_ADMIN_MSG}: {access_name};{path}')
             return 1
@@ -89,9 +103,12 @@ class Command(BaseCommand):
             return 1
         return 0
 
-    @classmethod
     def _check_coll_access(
-        cls, sample_path: str, admin_id: int, group_id: int, irods: iRODSSession
+        self,
+        sample_path: str,
+        admin_id: int,
+        group_id: int,
+        irods: iRODSSession,
     ) -> int:
         """
         Check collection ACLs under a project sample path.
@@ -102,12 +119,13 @@ class Command(BaseCommand):
         :param irods: iRODSSession object
         :return: Invalid access count (int)
         """
+        self._log_debug('Checking collection access..')
         ret = 0
         query = irods.query(
             Collection, CollectionAccess, CollectionUser
         ).filter(Like(Collection.name, sample_path + '%'))
         for r in query:
-            ret += cls._check_access(
+            ret += self._check_access(
                 r[CollectionAccess.user_id],
                 r[CollectionUser.name],
                 r[CollectionAccess.name],
@@ -116,11 +134,15 @@ class Command(BaseCommand):
                 group_id,
             )
         query.close()
+        self._log_debug(f'Collection check done, count={ret}')
         return ret
 
-    @classmethod
     def _check_obj_access(
-        cls, sample_path: str, admin_id: int, group_id: int, irods: iRODSSession
+        self,
+        sample_path: str,
+        admin_id: int,
+        group_id: int,
+        irods: iRODSSession,
     ) -> int:
         """
         Check data object ACLs under a project sample path.
@@ -131,13 +153,14 @@ class Command(BaseCommand):
         :param irods: iRODSSession object
         :return: Invalid access count (int)
         """
+        self._log_debug('Checking data object access..')
         ret = 0
         query = irods.query(DataObject, DataAccess, Collection, User).filter(
             Like(Collection.name, sample_path + '%')
         )
         for r in query:
             path = iRODSPath(r[Collection.name], r[DataObject.name])
-            ret += cls._check_access(
+            ret += self._check_access(
                 r[DataAccess.user_id],
                 r[User.name],
                 r[DataAccess.name],
@@ -146,6 +169,7 @@ class Command(BaseCommand):
                 group_id,
             )
         query.close()
+        self._log_debug(f'Data object check done, count={ret}')
         return ret
 
     def _check_project(
@@ -159,6 +183,7 @@ class Command(BaseCommand):
         :param irods: iRODSSession object
         :returns: Invalid access count (int)
         """
+        self._log_debug(f'Checking project {project.get_log_title()}..')
         sample_path = self.irods_backend.get_sample_path(project)
         project_group = self.irods_backend.get_group_name(project)
         group_id = irods.users.get(project_group).id
@@ -166,6 +191,7 @@ class Command(BaseCommand):
         ret = 0
         ret += self._check_coll_access(*check_args)
         ret += self._check_obj_access(*check_args)
+        self._log_debug(f'Project check done, count={ret}')
         return ret
 
     def handle(self, *args, **options):